Refactored to Stacktrace organization tree pattern
