+++ src/routes/index.routes.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const busRoutes = require('./bus.routes');
+const busRoutes = require('../bus/bus.routes');
 
 router.use('/bus', busRoutes);
